+++ RELEASE_NOTES.md
@@ -35,6 +35,8 @@ on the SWI-Prolog public forum.
 * ADDED: Lint check for redundant uses of the `{}/1` compiler bypass control
 construct (controlled by the `suspicious_calls` compiler flag).
 
+* ADDED: Lint check for cuts in clauses for multifile predicates.
+
 * ADDED: Lint check for possible non-steadfast predicate definitions.
 
 * ADDED: Compiler option `steadfastness` to control output of steadfast